feat: end-to-end pipeline smoke test with 83% accuracy on loan dataset
- Add __init__.py to all modules to register them as Python packages
- Implement Preprocessor class chaining Imputer, Scaler, and Encoder
- Fix numpy broadcasting bug: reshape y to (m,1) in LogisticRegression.fit
- Fix np.c_[] dtype bug: convert DataFrame to float64 before bias concat
- Fix Encoder reindex to enforce training column schema on test data
- Add score() method to LogisticRegression and LinearRegression
- Add pipeline_test.py: full DataManager → Preprocessor → Model test

+++ ml/preprocessing/preprocessor.py
@@ -1,6 +1,6 @@
-from preprocessing.imputer import Imputer
-from preprocessing.scaler import Scaler
-from preprocessing.encoder import Encoder
+from ml.preprocessing.imputer import Imputer
+from ml.preprocessing.scaler import Scaler
+from ml.preprocessing.encoder import Encoder
 class Preprocessor:
     
     def __init__(self,cat_cols,num_cols):
@@ -26,7 +26,7 @@ class Preprocessor:
         X = self.imputer.transform(X)
         X = self.scaler.transform(X)
         X = self.encoder.transform(X)
-        return X
+        return X.astype(float)
     
     def fit_transform(self,X):
         return self.fit(X).transform(X)
